# frontend/openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}},"/api/health":{"get":{"operationId":"AppController_getHealth","summary":"Health check endpoint","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}},"/api/api/auth/register":{"post":{"operationId":"AuthController_register","summary":"Register a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully"}},"tags":["auth"]}},"/api/api/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful"}},"tags":["auth"]}},"/api/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","summary":"Refresh access token","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Token refreshed successfully"}},"tags":["auth"]}},"/api/api/auth/verify":{"post":{"operationId":"AuthController_verify","summary":"Verify email or phone","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyDto"}}}},"responses":{"200":{"description":"Verification successful"}},"tags":["auth"]}},"/api/api/users/me":{"get":{"operationId":"UsersController_getProfile","summary":"Get current user profile","parameters":[],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_updateProfile","summary":"Update current user profile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/api/users/me/become-host":{"post":{"operationId":"UsersController_becomeHost","summary":"Become a host","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BecomeHostDto"}}}},"responses":{"201":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/api/users/{id}":{"get":{"operationId":"UsersController_findOne","summary":"Get user public profile","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/api/categories":{"post":{"operationId":"CategoriesController_create","summary":"Create category (admin only)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["categories"],"security":[{"bearer":[]}]},"get":{"operationId":"CategoriesController_findAll","summary":"List all categories","parameters":[],"responses":{"200":{"description":""}},"tags":["categories"]}},"/api/api/categories/{id}":{"get":{"operationId":"CategoriesController_findOne","summary":"Get category by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["categories"]},"patch":{"operationId":"CategoriesController_update","summary":"Update category (admin only)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["categories"],"security":[{"bearer":[]}]},"delete":{"operationId":"CategoriesController_remove","summary":"Delete category (admin only)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["categories"],"security":[{"bearer":[]}]}},"/api/api/listings":{"post":{"operationId":"ListingsController_create","summary":"Create a new listing (host only)","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateListingDto"}}}},"responses":{"201":{"description":""}},"tags":["listings"],"security":[{"bearer":[]}]},"get":{"operationId":"ListingsController_findAll","summary":"Search listings with filters","parameters":[{"name":"q","required":false,"in":"query","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"minPrice","required":false,"in":"query","schema":{"type":"number"}},{"name":"maxPrice","required":false,"in":"query","schema":{"type":"number"}},{"name":"lat","required":false,"in":"query","schema":{"type":"number"}},{"name":"lng","required":false,"in":"query","schema":{"type":"number"}},{"name":"radiusKm","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"availableFrom","required":false,"in":"query","schema":{"type":"string"}},{"name":"availableTo","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":20,"type":"number"}},{"name":"sortBy","required":false,"in":"query","schema":{"enum":["distance","price_asc","price_desc","date"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["listings"]}},"/api/api/listings/{id}":{"get":{"operationId":"ListingsController_findOne","summary":"Get listing details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["listings"]},"patch":{"operationId":"ListingsController_update","summary":"Update listing (host or admin)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateListingDto"}}}},"responses":{"200":{"description":""}},"tags":["listings"],"security":[{"bearer":[]}]},"delete":{"operationId":"ListingsController_remove","summary":"Delete listing (host or admin)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["listings"],"security":[{"bearer":[]}]}},"/api/api/bookings":{"post":{"operationId":"BookingsController_create","summary":"Create a new booking","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":""}},"tags":["bookings"],"security":[{"bearer":[]}]}},"/api/api/bookings/me":{"get":{"operationId":"BookingsController_findAll","summary":"Get user bookings","parameters":[],"responses":{"200":{"description":""}},"tags":["bookings"],"security":[{"bearer":[]}]}},"/api/api/bookings/{id}":{"get":{"operationId":"BookingsController_findOne","summary":"Get booking details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["bookings"],"security":[{"bearer":[]}]}},"/api/api/bookings/{id}/confirm":{"patch":{"operationId":"BookingsController_confirm","summary":"Confirm booking (host only)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["bookings"],"security":[{"bearer":[]}]}},"/api/api/bookings/{id}/pay":{"post":{"operationId":"BookingsController_pay","summary":"Simulate payment for booking","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PayBookingDto"}}}},"responses":{"201":{"description":""}},"tags":["bookings"],"security":[{"bearer":[]}]}},"/api/api/bookings/{id}/cancel":{"patch":{"operationId":"BookingsController_cancel","summary":"Cancel booking","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["bookings"],"security":[{"bearer":[]}]}},"/api/api/reviews":{"post":{"operationId":"ReviewsController_create","summary":"Create a review","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"tags":["reviews"],"security":[{"bearer":[]}]}},"/api/api/reviews/user/{userId}":{"get":{"operationId":"ReviewsController_findByUser","summary":"Get reviews for a user","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["reviews"]}},"/api/api/reviews/{id}":{"get":{"operationId":"ReviewsController_findOne","summary":"Get review details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["reviews"]}},"/api/api/admin/users":{"get":{"operationId":"AdminController_getAllUsers","summary":"Get all users (admin only)","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/api/admin/listings":{"get":{"operationId":"AdminController_getAllListings","summary":"Get all listings (admin only)","parameters":[],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/api/admin/flag":{"post":{"operationId":"AdminController_flagListing","summary":"Flag a listing for review (admin only)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FlagListingDto"}}}},"responses":{"201":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}},"/api/api/admin/logs":{"get":{"operationId":"AdminController_getLogs","summary":"Get admin action logs (admin only)","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["admin"],"security":[{"bearer":[]}]}}},"info":{"title":"RentEverything API","description":"Production-grade backend API for travel & vacation rental platform. Focus: Accommodation, Mobility, Water & Beach Activities.","version":"1.0","contact":{}},"tags":[{"name":"health","description":"Health check"},{"name":"auth","description":"Authentication endpoints"},{"name":"users","description":"User management"},{"name":"categories","description":"Category management"},{"name":"listings","description":"Listing management and search"},{"name":"bookings","description":"Booking management"},{"name":"reviews","description":"Review system"},{"name":"admin","description":"Admin operations"}],"servers":[{"url":"http://localhost:3000","description":"Development server"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"john@example.com"},"phone":{"type":"string","example":"+21612345678"},"password":{"type":"string","example":"password123","minLength":6}},"required":["name","password"]},"LoginDto":{"type":"object","properties":{"emailOrPhone":{"type":"string","example":"john@example.com"},"password":{"type":"string","example":"password123"}},"required":["emailOrPhone","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"VerifyDto":{"type":"object","properties":{"userId":{"type":"string"},"type":{"type":"string","enum":["email","phone"]},"code":{"type":"string"}},"required":["userId","type","code"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"avatarUrl":{"type":"string"}}},"BecomeHostDto":{"type":"object","properties":{"acceptTerms":{"type":"boolean","example":true}},"required":["acceptTerms"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string","example":"Bicycles"},"slug":{"type":"string","example":"bicycles"},"icon":{"type":"string","example":"🚲"},"allowed_for_private":{"type":"boolean","default":true}},"required":["name","allowed_for_private"]},"UpdateCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"slug":{"type":"string"},"icon":{"type":"string"},"allowed_for_private":{"type":"boolean"}}},"CreateListingDto":{"type":"object","properties":{"title":{"type":"string","example":"Mountain Bike for Rent"},"description":{"type":"string","example":"High-quality mountain bike..."},"categoryId":{"type":"string"},"pricePerDay":{"type":"number","example":25,"description":"Price per day in TND"},"latitude":{"type":"number","example":36.8475},"longitude":{"type":"number","example":11.0939},"address":{"type":"string","example":"123 Main St, Kelibia, Tunisia"},"rules":{"type":"string"},"availability":{"type":"array","items":{"type":"string"}}},"required":["title","description","categoryId","pricePerDay","latitude","longitude","address"]},"UpdateListingDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"pricePerDay":{"type":"number"},"latitude":{"type":"number"},"longitude":{"type":"number"},"address":{"type":"string"},"rules":{"type":"string"},"isActive":{"type":"boolean"}}},"CreateBookingDto":{"type":"object","properties":{"listingId":{"type":"string"},"startDate":{"type":"string","example":"2024-06-01"},"endDate":{"type":"string","example":"2024-06-05"}},"required":["listingId","startDate","endDate"]},"PayBookingDto":{"type":"object","properties":{"paymentToken":{"type":"string","description":"Payment token from payment gateway"},"receipt":{"type":"string","description":"Payment receipt URL or base64"}}},"CreateReviewDto":{"type":"object","properties":{"bookingId":{"type":"string"},"rating":{"type":"number","minimum":1,"maximum":5},"comment":{"type":"string"}},"required":["bookingId","rating"]},"FlagListingDto":{"type":"object","properties":{"listingId":{"type":"string"},"reason":{"type":"string","example":"Inappropriate content"}},"required":["listingId","reason"]}}}}